Add useAxiosFetch hook. Reconfigure Netlify function routes to '/api/*'
