+++ movie-app/src/hooks/useAxiosFetch.js
@@ -1,0 +1,33 @@
+import { useState, useEffect } from 'react';
+import axios from 'axios';
+
+function useAxiosFetch(url) {
+  const [data, setData] = useState([]);
+  const [error, setError] = useState(null);
+  const [isLoading, setIsLoading] = useState(false);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      setIsLoading(true);
+
+      try {
+        const { data: res } = await axios.get(url);
+
+        // Convert to JSON
+        // const jsonData = await res.json();
+
+        setData(res);
+      } catch (err) {
+        setError(err.message);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchData();
+  }, [url]);
+
+  return { data, error, isLoading };
+}
+
+export default useAxiosFetch;
