+++ movie-app/src/index.js
@@ -4,6 +4,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import App from './App';
 import Movies from './routes/movies';
 import TvShows from './routes/tvShows';
+import FetchMovie from './routes/api/fetch-movie';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
@@ -13,6 +14,7 @@ root.render(
         <Route path="movies" element={<Movies />} />
         <Route path="tvShows" element={<TvShows />} />
       </Route>
+      <Route path="/results/fetch-movie" element={<FetchMovie />} />
     </Routes>
   </BrowserRouter>,
 );
