+++ netlify.toml
@@ -1,2 +1,7 @@
 [build]
-    functions = "functions/"
+    functions = "functions"
+
+[[redirects]]
+    from = "/api/*"
+    to = "/.netlify/functions/:splat"
+    status = 200
